Add files via upload
Used the old VCARRS Jupyter Notebook to finally get my own version working. Filter parameters are that the mass needs to be between 10^10 and 10^12, and the b/a ratio needs to be between 0.2 and 0.5 (to prevent too much error from galaxies being too inclined and affecting the doppler shift). Calculated DM content for the first 100 galaxies that fit that filter condition, and plotted them at the very bottom. Average dark matter content seems to be roughly 5x the baryonic matter. Will widen the search soon!